+++ src/routes/challenges/$id/index.tsx
@@ -5,7 +5,7 @@ import { Button } from '@/components/shadcn/button'
 import { Label } from '@/components/shadcn/label'
 import { RadioGroup, RadioGroupItem } from '@/components/shadcn/radio-group'
 import { Separator } from '@/components/shadcn/separator'
-import { useChallenge, useChallengesParticipants } from '@/hooks/useChallenge'
+import { useChallenge, useChallengesParticipants } from '@/hooks/use-challenge'
 import { DataGrid, type GridColDef } from '@mui/x-data-grid'
 import { createFileRoute } from '@tanstack/react-router'
 import dayjs from 'dayjs'
