+++ src/routes/challenges/type/team.tsx
@@ -1,4 +1,4 @@
-import { useTeamChallenges } from '@/hooks/useChallenge'
+import { useTeamChallenges } from '@/hooks/use-challenge'
 import GlobalNavigation from '@/components/global-navigation'
 import PageContainer from '@/components/page-container'
 import PageTitle from '@/components/page-title'
